feat: 미션_코드 리펙토링
